Reconstruct network module.
