+++ network/src/p2p/peer.rs
@@ -5,10 +5,11 @@ use std::io::*;
 use std::result::Result as SerdeResult;
 
 use common::address::Address as Account;
-use message::protocol::*;
 use nat::*;
-use session::*;
-use session_state::*;
+
+use super::protocol::*;
+use super::session::*;
+use super::session_state::*;
 
 use mio::{Evented, Poll, PollOpt, Ready, Token};
 use mio::net::TcpStream;
